Signup and login page is completed
