+++ package.json
@@ -10,6 +10,7 @@
     "react": "^18.0.0",
     "react-bootstrap": "^2.2.2",
     "react-dom": "^18.0.0",
+    "react-geocode": "^0.2.3",
     "react-particles-js": "^3.6.0",
     "react-scripts": "5.0.0",
     "web-vitals": "^2.1.4",
