+++ src/App.js
@@ -1,13 +1,14 @@
 import logo from './logo.svg';
 import './App.css';
-import Signup from './components/Signup';
+//import SignupProvider from './components/SignupProvider';
+import SignupConsumer from './components/SignupConsumer';
 import LoginPage from "./components/LoginPage";
 import HomePage from './components/HomePage';
 
 function App() {
   return (
     <div className="App">
-      <Signup/>
+      <SignupConsumer/>
       <LoginPage />
       <HomePage/>
     </div>
